Finally correcting the test script
